refactor: 당첨 번호와 일치하는 개수 확인하는 메소드 리팩토링

+++ src/main/java/lotto/LottoTicket.java
@@ -8,7 +8,7 @@ public class LottoTicket {
         this.lottoNumbers = lottoNumbers;
     }
 
-    public int drawing(LottoNumbers winningNumbers) {
+    public long drawing(LottoNumbers winningNumbers) {
         return lottoNumbers.matchNumbers(winningNumbers);
     }
 }
